src/main/java/com/example/fanficapi/service/impl/UserDetailsServiceImpl.java
package com.example.fanficapi.service.impl;

import com.example.fanficapi.model.User;
import com.example.fanficapi.payload.UserDetailsImpl;
import com.example.fanficapi.service.UserService;
import lombok.RequiredArgsConstructor;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class UserDetailsServiceImpl implements UserDetailsService {

    private final UserService userService;

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        User user = userService.findByUsername(username);
        return UserDetailsImpl.build(user);
    }
}
